Astartransfer plus tests

+++ src/Algorithm/Heuristic.java
@@ -1,5 +1,5 @@
 package Algorithm;
 import Graph.*;
 public interface Heuristic {
-    double getHeuristicValue(Vertex current, Vertex end);
+    int getHeuristicValue(Vertex current, Vertex end);
 }
